(#110) Refactor: 프론트엔드 설정 코드
