+++ client/babel.config.json
@@ -3,7 +3,8 @@
     [
       "@babel/preset-env",
       {
-        "targets": "> 1%, not dead"
+        "targets": "> 1%, not dead",
+        "useBuiltIns": "usage"
       }
     ]
   ]
