+++ client/webpack.prod.js
@@ -7,7 +7,7 @@ export default merge(common, {
   mode: 'production',
   plugins: [
     new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('production'),
-    }),
+      'process.env.NODE_ENV': JSON.stringify('production')
+    })
   ]
 });
